uc-crux-llvm: Factor out Constraints into its own module
Will be renamed in the next commit.
