+++ uc-crux-llvm/src/UCCrux/LLVM/Run/Simulate.hs
@@ -105,7 +105,6 @@ import           Crux.LLVM.Simulate (setupSimCtxt)
  -- local
 import           UCCrux.LLVM.Classify (classifyAssertion, classifyBadBehavior)
 import           UCCrux.LLVM.Classify.Types (Located(Located), Explanation(..), Uncertainty(..))
-import           UCCrux.LLVM.Constraints (Constraints, returnConstraints, relationalConstraints)
 import           UCCrux.LLVM.Context.App (AppContext, log)
 import           UCCrux.LLVM.Context.Function (FunctionContext, functionName)
 import           UCCrux.LLVM.Context.Module (ModuleContext, llvmModule, moduleTranslation)
@@ -118,6 +117,7 @@ import           UCCrux.LLVM.Overrides.Polymorphic (PolymorphicLLVMOverride, get
 import           UCCrux.LLVM.Overrides.Unsound (UnsoundOverrideName, unsoundOverrides)
 import           UCCrux.LLVM.FullType.Type (FullType, MapToCrucibleType)
 import           UCCrux.LLVM.PP (ppRegMap, ppProgramLoc)
+import           UCCrux.LLVM.Precondition (Constraints, returnConstraints, relationalConstraints)
 import           UCCrux.LLVM.Run.Unsoundness (Unsoundness(Unsoundness))
 import           UCCrux.LLVM.Setup (setupExecution, SetupResult(SetupResult), SymValue)
 import           UCCrux.LLVM.Setup.Assume (assume)
