+++ uc-crux-llvm/test/Check.hs
@@ -29,12 +29,12 @@ import qualified Test.Tasty.HUnit as TH
 
 import qualified What4.Interface as What4
 
-import           UCCrux.LLVM.Constraints (emptyConstraints)
 import           UCCrux.LLVM.Context.Function (makeFunctionContext)
 import           UCCrux.LLVM.Context.Module (ModuleContext, CFGWithTypes(..), defnTypes, findFun, withModulePtrWidth)
 import           UCCrux.LLVM.Module (FuncSymbol(FuncDefnSymbol), DefnSymbol, defnSymbolToString)
 import           UCCrux.LLVM.Newtypes.FunctionName (functionNameFromString)
 import qualified UCCrux.LLVM.Overrides.Check as Check
+import           UCCrux.LLVM.Precondition (emptyConstraints)
 import           UCCrux.LLVM.Run.Check (SomeCheckResult(SomeCheckResult), inferThenCheck, getCheckResult, getSomeCheckedCalls)
 import qualified UCCrux.LLVM.Run.EntryPoints as EntryPoints
 import           UCCrux.LLVM.Run.Loop (bugfindingLoop)
